Removed all the code about Thumbnailer timer. (The render function that generates the thumbnails has to run in the GUI thread to render the page. If this function is driven by a timer, reentrances of it will happen, which may make nixnote crash. This commit makes the thumbnails only generated when browsering or saving a note.)
Removed the thumbnailer member variables defined in nixnote.cpp and nbrowserwindow.cpp but not used.
Left only one thumbnailer, the hammer.

Made the thumbnailing Enable/Disable as the Tableview header shows/hides the thumbnail column.

Made the calculation of the thumbnailers in statistics depend on the number of the files under
global.fileManager.getThumbnailDirPath.

Disabled thumbnailing by default because it may slow down the GUI thread response speed.